Make parsing sturdier
Before this change we had to hand-patch because not everything would've
propagated to the resulting binary.

The solution is two-fold:

 - Rely as less as possible on assumptions about the AST (such as what
children the current node has) through a more extensive use of recursion

 - Accumulate pointers and append them to the next non-pointer node,
except for function pointers (which we handle differently because C)

Now a wrapper for `wayland-client-core` v1.21.0 doesn't need any patch!